moved load_player_names into dependencies.py
