+++ civ_vi_webhook/dependencies.py
@@ -13,3 +13,14 @@ def load_most_recent_games() -> dict:
     except FileNotFoundError:
         api_logger.warning("Prior JSON file not found. If this is your first run, this is OK.")
     return games
+
+
+def load_player_names() -> dict:
+    """If player names have been defined, load them in."""
+    try:
+        with open('player_names.conf', 'r') as file:
+            player_names = json.load(file)
+            api_logger.debug("Player Conversion file loaded.")
+    except FileNotFoundError:
+        api_logger.warning("No Player Conversion file loaded. Messages will use Steam account names.")
+    return player_names
